requirements-windows.txt
asgiref==3.7.2
Django==3.2.25
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.0
Pillow==9.5.0
psycopg2-binary==2.9.9
PyJWT==1.7.1
python-decouple==3.8
pytz==2024.1
sqlparse==0.4.4
typing-extensions==4.7.1
